+++ WelsonJS.Augmented/WelsonJS.Launcher/ApiEndpoints/DnsQuery.cs
@@ -13,7 +13,7 @@ using System.Net.Sockets;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace WelsonJS.Launcher.ResourceTools
+namespace WelsonJS.Launcher.ApiEndpoints
 {
     public class DnsQuery : IApiEndpoint
     {
