+++ WelsonJS.Augmented/WelsonJS.Launcher/ApiEndpoints/Whois.cs
@@ -10,7 +10,7 @@ using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace WelsonJS.Launcher.ResourceTools
+namespace WelsonJS.Launcher.ApiEndpoints
 {
     public class Whois : IApiEndpoint
     {
